add controllers for contacts
