+++ controllers/contacts/getUserListContacts.js
@@ -1,0 +1,10 @@
+const { Contact } = require("../../models");
+const { sendSuccessResponse } = require("../../utils");
+
+const getUserListContacts = async (req, res) => {
+  const { _id } = req.user;
+  const contacts = await Contact.find({ owner: _id });
+  sendSuccessResponse(res, { contacts });
+};
+
+module.exports = getUserListContacts;
