+++ controllers/contacts/removeContactById.js
@@ -1,0 +1,16 @@
+const { NotFound } = require("http-errors");
+const { Contact } = require("../../models");
+const { sendSuccessResponse } = require("../../utils");
+
+const removeContactById = async (req, res) => {
+  const { contactId } = req.params;
+  const contact = await Contact.findByIdAndRemove(contactId);
+
+  if (!contact) {
+    throw new NotFound(`Contact with id=${contactId} not found`);
+  }
+
+  sendSuccessResponse(res, { message: "Success remove", contact });
+};
+
+module.exports = removeContactById;
