+++ controllers/contacts/updateContactById.js
@@ -1,0 +1,16 @@
+const { NotFound } = require("http-errors");
+const { Contact } = require("../../models");
+const { sendSuccessResponse } = require("../../utils");
+
+const updateContactById = async (req, res) => {
+  const { contactId } = req.params;
+  const contact = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
+  if (!contact) {
+    throw new NotFound(`Contact with id=${contactId} not found`);
+  }
+  sendSuccessResponse(res, { contact });
+};
+
+module.exports = updateContactById;
