+++ controllers/index.js
@@ -1,5 +1,7 @@
 const contacts = require("./contacts");
+const auth = require("./auth");
 
 module.exports = {
   contacts,
+  auth,
 };
